services/insert_index/pyproject.toml
[project]
name = "insert-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-servicebus",
    "azure-storage-blob",
    "fastapi",
    "hnswlib",
    "numpy",
    "pydantic",
    "python-dotenv",
    "scikit-learn",
    "uvicorn",
]
